kernel: split system.{c,h}, add systime.{c,h}, sysstr.{c,h}
Also get rid of SySleep and SyUSleep, which were direct wrappers
around sleep() and usleep().
