+++ Makefile.rules
@@ -98,7 +98,9 @@ SOURCES += src/stringobj.c
 SOURCES += src/syntaxtree.c
 SOURCES += src/sysfiles.c
 SOURCES += src/sysmem.c
+SOURCES += src/sysstr.c
 SOURCES += src/system.c
+SOURCES += src/systime.c
 SOURCES += src/tietze.c
 SOURCES += src/tracing.c
 SOURCES += src/trans.cc
