+++ src/scanner.h
@@ -17,6 +17,7 @@
 #ifndef GAP_SCANNER_H
 #define GAP_SCANNER_H
 
+#include "sysstr.h"
 #include "system.h"
 
 /****************************************************************************
