+++ src/system.c
@@ -23,6 +23,7 @@
 #include "sysfiles.h"
 #include "sysmem.h"
 #include "sysopt.h"
+#include "sysstr.h"
 
 #ifdef HPCGAP
 #include "hpc/misc.h"
@@ -43,17 +44,6 @@
 
 #ifdef HAVE_LIBREADLINE
 #include <readline/readline.h>
-#endif
-
-#include <sys/time.h>                   /* definition of 'struct timeval' */
-#include <sys/types.h>
-
-#ifdef HAVE_SYS_RESOURCE_H
-#include <sys/resource.h>               /* definition of 'struct rusage' */
-#endif
-
-#ifdef SYS_IS_DARWIN
-#include <mach/mach_time.h>
 #endif
 
 
@@ -285,186 +275,6 @@ UInt SyWindow;
 
 /****************************************************************************
 **
-*F * * * * * * * * * * * * * time related functions * * * * * * * * * * * * *
-*/
-
-/****************************************************************************
-**
-*F  SyTime()  . . . . . . . . . . . . . . . return time spent in milliseconds
-**
-**  'SyTime' returns the number of milliseconds spent by GAP so far.
-**
-**  Should be as accurate as possible,  because it  is  used  for  profiling.
-*/
-UInt SyTime ( void )
-{
-    struct rusage       buf;
-
-    if ( getrusage( RUSAGE_SELF, &buf ) ) {
-        Panic("'SyTime' could not get time");
-    }
-    return buf.ru_utime.tv_sec*1000 + buf.ru_utime.tv_usec/1000;
-}
-UInt SyTimeSys ( void )
-{
-    struct rusage       buf;
-
-    if ( getrusage( RUSAGE_SELF, &buf ) ) {
-        Panic("'SyTimeSys' could not get time");
-    }
-    return buf.ru_stime.tv_sec*1000 + buf.ru_stime.tv_usec/1000;
-}
-UInt SyTimeChildren ( void )
-{
-    struct rusage       buf;
-
-    if ( getrusage( RUSAGE_CHILDREN, &buf ) ) {
-        Panic("'SyTimeChildren' could not get time");
-    }
-    return buf.ru_utime.tv_sec*1000 + buf.ru_utime.tv_usec/1000;
-}
-UInt SyTimeChildrenSys ( void )
-{
-    struct rusage       buf;
-
-    if ( getrusage( RUSAGE_CHILDREN, &buf ) ) {
-        Panic("'SyTimeChildrenSys' could not get time");
-    }
-    return buf.ru_stime.tv_sec*1000 + buf.ru_stime.tv_usec/1000;
-}
-
-
-/****************************************************************************
-**
-*F * * * * * * * * * * * * * * * string functions * * * * * * * * * * * * * *
-*/
-
-
-#ifndef HAVE_STRLCPY
-
-size_t strlcpy (
-    char *dst,
-    const char *src,
-    size_t len)
-{
-    /* Keep a copy of the original src. */
-    const char * const orig_src = src;
-
-    /* If a non-empty len was specified, we can actually copy some data. */
-    if (len > 0) {
-        /* Copy up to len-1 bytes (reserve one for the terminating zero). */
-        while (--len > 0) {
-            /* Copy from src to dst; if we reach the string end, we are
-               done and can simply return the total source string length */
-            if ((*dst++ = *src++) == 0) {
-                /* return length of source string without the zero byte */
-                return src - orig_src - 1;
-            }
-        }
-        
-        /* If we got here, then we used up the whole buffer and len is zero.
-           We must make sure to terminate the destination string. */
-        *dst = 0;
-    }
-
-    /* in the end, we must return the length of the source string, no
-       matter whether we completely copied or not; so advance src
-       till its terminator is reached */
-    while (*src++)
-        ;
-
-    /* return length of source string without the zero byte */
-    return src - orig_src - 1;
-}
-
-#endif /* !HAVE_STRLCPY */
-
-
-#ifndef HAVE_STRLCAT
-
-size_t strlcat (
-    char *dst,
-    const char *src,
-    size_t len)
-{
-    /* Keep a copy of the original dst. */
-    const char * const orig_dst = dst;
-
-    /* Find the end of the dst string, so that we can append after it. */
-    while (*dst != 0 && len > 0) {
-        dst++;
-        len--;
-    }
-
-    /* We can only append anything if there is free space left in the
-       destination buffer. */
-    if (len > 0) {
-        /* One byte goes away for the terminating zero. */
-        len--;
-
-        /* Do the actual work and append from src to dst, until we either
-           appended everything, or reached the dst buffer's end. */
-        while (*src != 0 && len > 0) {
-            *dst++ = *src++;
-            len--;
-        }
-
-        /* Terminate, terminate, terminate! */
-        *dst = 0;
-    }
-
-    /* Compute the final result. */
-    return (dst - orig_dst) + strlen(src);
-}
-
-#endif /* !HAVE_STRLCAT */
-
-size_t strxcpy (
-    char *dst,
-    const char *src,
-    size_t len)
-{
-    size_t res = strlcpy(dst, src, len);
-    assert(res < len);
-    return res;
-}
-
-size_t strxcat (
-    char *dst,
-    const char *src,
-    size_t len)
-{
-    size_t res = strlcat(dst, src, len);
-    assert(res < len);
-    return res;
-}
-
-/****************************************************************************
-**
-*F  SySleep( <secs> ) . . . . . . . . . . . . . .sleep GAP for <secs> seconds
-**
-**  NB Various OS events (like signals) might wake us up
-**
-*/
-void SySleep ( UInt secs )
-{
-  sleep( (unsigned int) secs );
-}
-
-/****************************************************************************
-**
-*F  SyUSleep( <msecs> ) . . . . . . . . . .sleep GAP for <msecs> microseconds
-**
-**  NB Various OS events (like signals) might wake us up
-**
-*/
-void SyUSleep ( UInt msecs )
-{
-  usleep( (unsigned int) msecs );
-}
-
-/****************************************************************************
-**
 *F * * * * * * * * * * * * * initialize module * * * * * * * * * * * * * * *
 */
 
@@ -506,141 +316,6 @@ void Panic_(const char * file, int line, const char * fmt, ...)
     va_end (args);
     fputs("\n", stderr);
     SyExit(1);
-}
-
-
-/****************************************************************************
-**
-*F  SyNanosecondsSinceEpoch()
-**
-**  'SyNanosecondsSinceEpoch' returns a 64-bit integer which represents the
-**  number of nanoseconds since some unspecified starting point. This means
-**  that the number returned by this function is not in itself meaningful,
-**  but the difference between the values returned by two consecutive calls
-**  can be used to measure wallclock time.
-**
-**  The accuracy of this is system dependent. For systems that implement
-**  clock_getres, we could get the promised accuracy.
-**
-**  Note that gettimeofday has been marked obsolete in the POSIX standard.
-**  We are using it because it is implemented in most systems still.
-**
-**  If we are using gettimeofday we cannot guarantee the values that
-**  are returned by SyNanosecondsSinceEpoch to be monotonic.
-**
-**  Returns -1 to represent failure
-**
-*/
-Int8 SyNanosecondsSinceEpoch(void)
-{
-  Int8 res;
-
-#if defined(SYS_IS_DARWIN)
-  static mach_timebase_info_data_t timeinfo;
-  if ( timeinfo.denom == 0 ) {
-    (void) mach_timebase_info(&timeinfo);
-  }
-  res = mach_absolute_time();
-
-  res *= timeinfo.numer;
-  res /= timeinfo.denom;
-#elif defined(HAVE_CLOCK_GETTIME) && defined(CLOCK_MONOTONIC)
-  struct timespec ts;
-
-  if (clock_gettime(CLOCK_MONOTONIC, &ts) == 0) {
-    res = ts.tv_sec;
-    res *= 1000000000L;
-    res += ts.tv_nsec;
-  } else {
-    res = -1;
-  }
-#elif defined(HAVE_GETTIMEOFDAY)
-  struct timeval tv;
-
-  if (gettimeofday(&tv, NULL) == 0) {
-    res = tv.tv_sec;
-    res *= 1000000L;
-    res += tv.tv_usec;
-    res *= 1000;
-  } else {
-    res = -1;
-  };
-#else
-  res = -1;
-#endif
-
-  return res;
-}
-
-
-/****************************************************************************
-**
-*V  SyNanosecondsSinceEpochMethod
-*V  SyNanosecondsSinceEpochMonotonic
-**  
-**  These constants give information about the method used to obtain
-**  NanosecondsSinceEpoch, and whether the values returned are guaranteed
-**  to be monotonic.
-*/
-#if defined(SYS_IS_DARWIN)
-  const char * const SyNanosecondsSinceEpochMethod = "mach_absolute_time";
-  const Int SyNanosecondsSinceEpochMonotonic = 1;
-#elif defined(HAVE_CLOCK_GETTIME) && defined(CLOCK_MONOTONIC)
-  const char * const SyNanosecondsSinceEpochMethod = "clock_gettime";
-  const Int SyNanosecondsSinceEpochMonotonic = 1;
-#elif defined(HAVE_GETTIMEOFDAY)
-  const char * const SyNanosecondsSinceEpochMethod = "gettimeofday";
-  const Int SyNanosecondsSinceEpochMonotonic = 0;
-#else
-  const char * const SyNanosecondsSinceEpochMethod = "unsupported";
-  const Int SyNanosecondsSinceEpochMonotonic = 0;
-#endif
-
-
-/****************************************************************************
-**
-*F  SyNanosecondsSinceEpochResolution()
-**
-**  'SyNanosecondsSinceEpochResolution' returns a 64-bit integer which
-**  represents the resolution in nanoseconds of the timer used for
-**  SyNanosecondsSinceEpoch. 
-**
-**  If the return value is positive then the value has been returned
-**  by the operating system can can probably be relied on. If the 
-**  return value is negative it is just an estimate (as in the case
-**  of gettimeofday we have no way to get the exact resolution so we
-**  just pretend that the resolution is 1000 nanoseconds).
-**
-**  A result of 0 signifies inability to obtain any sensible value.
-*/
-Int8 SyNanosecondsSinceEpochResolution(void)
-{
-  Int8 res;
-
-#if defined(SYS_IS_DARWIN)
-  static mach_timebase_info_data_t timeinfo;
-  if ( timeinfo.denom == 0 ) {
-    (void) mach_timebase_info(&timeinfo);
-  }
-  res = timeinfo.numer;
-  res /= timeinfo.denom;
-#elif defined(HAVE_CLOCK_GETTIME) && defined(CLOCK_MONOTONIC)
-  struct timespec ts;
-
-  if (clock_getres(CLOCK_MONOTONIC, &ts) == 0) {
-    res = ts.tv_sec;
-    res *= 1000000000L;
-    res += ts.tv_nsec;
-  } else {
-    res = 0;
-  }
-#elif defined(HAVE_GETTIMEOFDAY)
-  res = -1000;
-#else
-  res = 0;
-#endif
-
-  return res;
 }
 
 /****************************************************************************
