+++ src/system.h
@@ -269,126 +269,7 @@ enum { BIPEB = sizeof(UInt) * 8, LBIPEB = (BIPEB == 64) ? 6 : 5 };
 
 /****************************************************************************
 **
-*F * * * * * * * * * * * * * time related functions * * * * * * * * * * * * *
 */
-
-/****************************************************************************
-**
-*F  SyTime()  . . . . . . . . . . . . . . . return time spent in milliseconds
-**
-**  'SyTime' returns the number of milliseconds spent by GAP so far.
-**
-**  Should be as accurate as possible,  because it  is  used  for  profiling.
-*/
-UInt SyTime(void);
-
-/* TODO: Properly document the following three calls */
-UInt SyTimeSys(void);
-UInt SyTimeChildren(void);
-UInt SyTimeChildrenSys(void);
-
-/****************************************************************************
-**
-*F * * * * * * * * * * * * * * * string functions * * * * * * * * * * * * * *
-*/
-
-
-/****************************************************************************
-**
-*F  IsAlpha( <ch> ) . . . . . . . . . . . . .  is a character a normal letter
-**
-**  'IsAlpha' returns 1 if its character argument is a normal character  from
-**  the range 'a..zA..Z' and 0 otherwise.
-*/
-#define IsAlpha(ch)     (isalpha((unsigned int)ch))
-
-
-/****************************************************************************
-**
-*F  IsDigit( <ch> ) . . . . . . . . . . . . . . . . .  is a character a digit
-**
-**  'IsDigit' returns 1 if its character argument is a digit from  the  range
-**  '0..9' and 0 otherwise.
-*/
-#define IsDigit(ch)     (isdigit((unsigned int)ch))
-
-
-/****************************************************************************
-**
-*F  strlcpy( <dst>, <src>, <len> )
-**
-**  Copy <src> to buffer <dst> of size <len>. At most <len>-1 characters will
-**  be copied. Afterwards, <dst> is always 'NUL' terminated
-**  (unless <len> == 0).
-**
-**  Returns 'strlen( <src> )'; hence if the return value is greater or equal
-**  than <len>, truncation occurred.
-**
-**  This function is provided by some systems (e.g. OpenBSD, Mac OS X),
-**  but not by all, so we provide a fallback implementation for those
-**  systems that lack it.
-*/
-#ifndef HAVE_STRLCPY
-size_t strlcpy (
-    char *dst,
-    const char *src,
-    size_t len);
-#endif
-
-/****************************************************************************
-**
-*F  strlcat( <dst>, <src>, <len> )
-**
-**  Appends <src> to buffer <dst> of size <len> (unlike 'strncat', <len> is
-**  the full size of <dst>, not space left).
-**  At most <len>-1 characters will be copied.
-**  Afterwards, <dst> is always 'NUL' terminated (unless <len> == 0).
-**
-**  Returns initial length of <dst> plus 'strlen(<src>)'; hence if the return
-**  value is greater or equal than <len>, truncation occurred.
-**
-**  This function is provided by some systems (e.g. OpenBSD, Mac OS X),
-**  but not by all, so we provide a fallback implementation for those
-**  systems that lack it.
-*/
-#ifndef HAVE_STRLCAT
-size_t strlcat (
-    char *dst,
-    const char *src,
-    size_t len);
-#endif
-
-/****************************************************************************
-**
-*F  strxcpy( <dst>, <src>, <len> )
-**
-**  Copy <src> to buffer <dst> of size <len>.
-**  If an overflow would occur, trigger an assertion.
-**
-**  This should be used with caution; in general, proper error handling is
-**  preferable.
-**/
-size_t strxcpy (
-    char *dst,
-    const char *src,
-    size_t len);
-
-/****************************************************************************
-**
-*F  strxcat( <dst>, <src>, <len> )
-**
-**  Append <src> to buffer <dst> of size <len>.
-**  If an overflow would occur, trigger an assertion.
-**
-**  This should be used with caution; in general, proper error handling is
-**  preferable.
-**/
-size_t strxcat (
-    char *dst,
-    const char *src,
-    size_t len);
-
-
 typedef const struct init_info StructInitInfo;
 typedef StructInitInfo* (*InitInfoFunc)(void);
 
@@ -422,52 +303,6 @@ void Panic_(const char * file, int line, const char * fmt, ...) NORETURN
 #define Panic(...) \
     Panic_(__FILE__, __LINE__, __VA_ARGS__)
 
-
-/****************************************************************************
-**
-*F  SyNanosecondsSinceEpoch()
-**
-**  'SyNanosecondsSinceEpoch' returns a 64-bit integer which represents the
-**  number of nanoseconds since some unspecified starting point. This means
-**  that the number returned by this function is not in itself meaningful,
-**  but the difference between the values returned by two consecutive calls
-**  can be used to measure wallclock time.
-**
-**  The accuracy of this is system dependent. For systems that implement
-**  clock_getres, we could get the promised accuracy.
-**
-**  Note that gettimeofday has been marked obsolete in the POSIX standard.
-**  We are using it because it is implemented in most systems still.
-**
-**  If we are using gettimeofday we cannot guarantee the values that
-**  are returned by SyNanosecondsSinceEpoch to be monotonic.
-**
-**  Returns -1 to represent failure
-**
-*/
-Int8 SyNanosecondsSinceEpoch(void);
-Int8 SyNanosecondsSinceEpochResolution(void);
-
-extern const char * const SyNanosecondsSinceEpochMethod;
-extern const Int SyNanosecondsSinceEpochMonotonic;
-
-/****************************************************************************
-**
-*F  SySleep( <secs> ) . . . . . . . . . . . . Try to sleep for <secs> seconds
-**
-**  The OS may wake us earlier, for example on receipt of a signal
-*/
-
-void SySleep(UInt secs);
-
-/****************************************************************************
-**
-*F  SyUSleep( <msecs> ) . . . . . . . . .Try to sleep for <msecs> microseconds
-**
-**  The OS may wake us earlier, for example on receipt of a signal
-*/
-
-void SyUSleep(UInt msecs);
 
 /****************************************************************************
 **
